make wallet connection a service
